Move crypto provider logic into SecureStorage.
